refactor event handlers
