+++ CHANGELOG.md
@@ -9,5 +9,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support opening a related url in view mode [#7](https://github.com/dpickett/open-related-url/issues/7)
+- Open a url in a link context menu [#6](https://github.com/dpickett/open-related-url/issues/6)
 
 [Unreleased]: https://github.com/dpickett/open-related-url/compare/0.0.11...HEAD
