+++ src/open-related-url/OpenRelatedUrlPlugin.ts
@@ -17,27 +17,35 @@ import {
 import SettingTab from "./SettingTab";
 import openUrl from "src/openUrl";
 import { LinkMenuEvent } from "./LinkMenuEvent";
+import { FileOpenEvent } from "./FileOpenEvent";
 
 export default class OpenRelatedUrlPlugin extends Plugin {
   settings: OpenRelatedUrlPluginSettings;
   linkMenuEvent: LinkMenuEvent;
+  fileOpenEvent: FileOpenEvent;
 
   async onload() {
     await this.loadSettings();
-    this.registerEvent(
-      this.app.workspace.on("file-open", (file: TFile) => {
-        this.registerCommands(file);
-      })
-    );
+
     if (!this.linkMenuEvent) {
       this.linkMenuEvent = new LinkMenuEvent(this);
       this.linkMenuEvent.registerEvent();
     }
+
+    if (!this.fileOpenEvent) {
+      this.fileOpenEvent = new FileOpenEvent(this);
+      this.fileOpenEvent.registerEvent();
+    }
+
+    this.addSettingTab(new SettingTab(this.app, this));
   }
 
   onunload() {
     if (this.linkMenuEvent) {
       this.linkMenuEvent.unregisterEvent();
+    }
+    if (this.fileOpenEvent) {
+      this.fileOpenEvent.unregisterEvent();
     }
   }
 
@@ -47,49 +55,5 @@ export default class OpenRelatedUrlPlugin extends Plugin {
 
   async saveSettings() {
     await this.saveData(this.settings);
-  }
-
-  private async registerCommands(file: TFile) {
-    const { urlFrontMatterNameSuffix } = this.settings;
-    // This adds an editor command that can perform some operation on the current editor instance
-    this.addCommand({
-      id: "open-related-url",
-      name: "Open Related URL",
-      callback: () => {
-        const frontMatter = resolveFrontMatter(app.metadataCache, file);
-        if (frontMatter) {
-          const urlSet = extractUrlSet(frontMatter, {
-            urlFrontMatterNameSuffix: urlFrontMatterNameSuffix,
-          });
-          new UrlModal(this.app, urlSet).open();
-        }
-      },
-    });
-
-    this.settings.quickNavigateNames.forEach((name) => {
-      this.addCommand({
-        id: `open-quick-url-${name}`,
-        name: `Quick Nav - ${name}`,
-        callback: () => {
-          const frontMatter = resolveFrontMatter(app.metadataCache, file);
-
-          let urlItem;
-          if (frontMatter) {
-            const urlSet = extractUrlSet(frontMatter, {
-              urlFrontMatterNameSuffix: urlFrontMatterNameSuffix,
-            });
-            urlItem = urlSet.find((url) => url.name === name);
-          }
-
-          if (urlItem) {
-            openUrl(urlItem.url);
-          } else {
-            new Notice(`URL for ${name} not found in frontmatter`);
-          }
-        },
-      });
-    });
-
-    this.addSettingTab(new SettingTab(this.app, this));
   }
 }
